fix(constraints): Mutating now strictly respects max size constraint
There was a where the minimum size (or no size) of an object was being used to calculate how much free space can be used when constructing or mutating a dynamic-sized object. Instead we now take the following approach for constructing an item:

1. Calculate the *max* size of the object (i.e. the largest enum variant) with the *minimum* non-zero number of items for dynamic-sized items. This is the minimum number of bytes required to conform to the min constraints while expressing every variant.
2. Call `new_fuzzed` on each field with the `max_size` constraint of BufferMaxSize - StructSize in random order to try to grant some fairness.
3. If a field consumed *less* space than expected, give the delta back for the remainder of the fields.

And when mutating:

1. Calculate the BufferMaxSize - ObjectCurrentSize.
2. Provide this size as a constraint when calling `mutate()`

There were a number of bugs specifically in mutating vecs, slices, and arrays where no max size was provided at all, which could result in extremely large object sizes.